Make `undo` and `redo` take `&mut self`
This is needed where the meaning of an action changes with the
modification, such as an allowed range in a list to insert points.
Inserting a point increases the allowed range with 1, since it refers
to the initial range in the list.

Therefore, `undo` and `redo` must be able to mutate the modifier.